x86 ported

+++ kernel/platform/x86/arch/common/include/ewokos_config.h
@@ -3,7 +3,7 @@
 
 #include <stdint.h>
 
-typedef uint32_t ewokos_addr_t;
+typedef uint64_t ewokos_addr_t;
 
 #define EWOK_STACK_ALIGN      8U
 #define EWOK_STACK_INIT_BIAS  8U
